+++ test/data_processing/test_data_processing.py
@@ -14,10 +14,14 @@
 
 # pylint: disable=unbalanced-tuple-unpacking
 
-from test.data_processing.fake_experiment import FakeExperiment, BaseDataProcessorTest
+from test.fake_experiment import FakeExperiment
+
 import numpy as np
 from qiskit.result.models import ExperimentResultData, ExperimentResult
 from qiskit.result import Result
+
+from qiskit.test import QiskitTestCase
+from qiskit.qobj.common import QobjExperimentHeader
 
 from qiskit_experiments import ExperimentData
 from qiskit_experiments.data_processing.data_processor import DataProcessor
@@ -30,6 +34,27 @@ from qiskit_experiments.data_processing.nodes import (
     Probability,
     MinMaxNormalize,
 )
+
+
+class BaseDataProcessorTest(QiskitTestCase):
+    """Define some basic setup functionality for data processor tests."""
+
+    def setUp(self):
+        """Define variables needed for most tests."""
+        super().setUp()
+
+        self.base_result_args = dict(
+            backend_name="test_backend",
+            backend_version="1.0.0",
+            qobj_id="id-123",
+            job_id="job-123",
+            success=True,
+        )
+
+        self.header = QobjExperimentHeader(
+            memory_slots=2,
+            metadata={"experiment_type": "fake_test_experiment"},
+        )
 
 
 class DataProcessorTest(BaseDataProcessorTest):
